air_revitalization/requirements.txt
nats-py>=2.7.0
asyncio-nats-client>=0.11.4
